feat(fips): Add FIPS compliance option for aws Hypershift clusters
Adds a new `fips` boolean flag to the chart to allow provisioning FIPS-compliant Hypershift clusters on AWS.

When `fips` is set to `true`, the `hypershift create cluster aws` command will be invoked with the `--fips` flag. This configures the hosted OpenShift cluster to operate in FIPS mode, ensuring:
- All cryptographic operations use FIPS 140-2 (or 140-3, if applicable) validated modules.
- The underlying RHCOS nodes are FIPS-enabled.
- Network communication utilizes FIPS-compliant protocols, specifically by setting `networkType: OVNKubernetes` and enabling IPSec.

This feature is crucial for environments requiring strict security and regulatory compliance.

Reference:
- Hypershift FIPS flag: https://github.com/openshift/hypershift/blob/9720e90964cf527422033550889835ee450ba46a/cmd/cluster/core/create.go#L88

Signed-off-by: Hasan Awad <hasan.m.awad94@gmail.com>